+++ lib/src/utils/config-reader/properties_config_reader.dart
@@ -18,6 +18,24 @@ class PropertiesConfigReader implements ConfigReader {
   }
 
   @override
+  List<String>? getList(String key, {String? deprecatedKey}) {
+    return get(key, deprecatedKey, (key) {
+      final value = _properties.get(key);
+      if (value != null && value.isNotEmpty) {
+        final stripped = value.startsWith('[') && value.endsWith(']')
+            ? value.substring(1, value.length - 1)
+            : value;
+        return stripped
+            .split(',')
+            .map((e) => e.trim())
+            .where((e) => e.isNotEmpty)
+            .toList();
+      }
+      return null;
+    });
+  }
+
+  @override
   bool contains(String key) {
     return _properties.contains(key);
   }