+++ test/configuration_values_test.dart
@@ -36,7 +36,8 @@ void main() {
         "--sentry-define=bin_dir=fixture-bin_dir",
         "--sentry-define=sentry_cli_cdn_url=fixture-sentry_cli_cdn_url",
         "--sentry-define=sentry_cli_version=1.0.0",
-        "--sentry-define=legacy_web_symbolication=true"
+        "--sentry-define=legacy_web_symbolication=true",
+        "--sentry-define=ignore_web_source_paths=some/, **/*.js"
       ];
       final sut = ConfigurationValues.fromArguments(arguments);
       expect(sut.name, 'fixture-sentry-name');
@@ -62,6 +63,7 @@ void main() {
       expect(sut.sentryCliCdnUrl, 'fixture-sentry_cli_cdn_url');
       expect(sut.sentryCliVersion, '1.0.0');
       expect(sut.legacyWebSymbolication, isTrue);
+      expect(sut.ignoreWebSourcePaths, ['some/', '**/*.js']);
     });
 
     test("fromArguments supports deprecated fields", () {
@@ -106,6 +108,9 @@ void main() {
       sentry_cli_cdn_url: fixture-sentry_cli_cdn_url
       sentry_cli_version: 1.0.0
       legacy_web_symbolication: true
+      ignore_web_source_paths:
+        - some/
+        - '**/*.js'
       ''';
 
       FileSystem fs = MemoryFileSystem.test();
@@ -149,6 +154,7 @@ void main() {
       expect(sut.binDir, 'fixture-bin_dir');
       expect(sut.sentryCliCdnUrl, 'fixture-sentry_cli_cdn_url');
       expect(sut.legacyWebSymbolication, isTrue);
+      expect(sut.ignoreWebSourcePaths, ['some/', '**/*.js']);
     });
 
     test('from config reader as properties', () {
@@ -172,6 +178,7 @@ void main() {
       bin_dir=fixture-bin_dir
       sentry_cli_cdn_url=fixture-sentry_cli_cdn_url
       sentry_cli_version=1.0.0
+      ignore_web_source_paths=[some/, **/*.js]
       ''';
 
       FileSystem fs = MemoryFileSystem.test();
@@ -215,6 +222,7 @@ void main() {
       expect(sut.binDir, 'fixture-bin_dir');
       expect(sut.sentryCliCdnUrl, 'fixture-sentry_cli_cdn_url');
       expect(sut.sentryCliVersion, '1.0.0');
+      expect(sut.ignoreWebSourcePaths, ['some/', '**/*.js']);
     });
 
     test('from config reader pubspec & properties', () {